reordered and renamed to org.ui.dialogs.SimpleDialog...
